Move toward big integer (2^64) (#1)
* update helper

* more updates

* correctness fixes

* update add funciton

* update subtract

* update subInPlace

* hnn

* finalize

+++ src/types.h
@@ -3,7 +3,7 @@
 
 #include <vector>
 
-using ByteArray = std::vector<std::uint8_t>;
+using ByteArray = std::vector<std::uint64_t>;
 constexpr int INVALID_START_BIT_INDEX = -1;
 
 #endif  // RSA_ENCRYPTOR_TYPES_H
